refactor: theme, responsive, getit.
